# api/openapi/rest.yaml
openapi: 3.0.3
info:
  title: TimeLog API
  version: 1.0.0
  description: |
    REST API for the TimeLog application. Payload DTO schemas are generated from
    `api/proto/timelog/v1`. The `{ data, message, status }` envelope is defined here.
servers:
  - url: /api
    description: API base path (relative to app origin)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ApiError:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                nullable: true
              message:
                type: string
              status:
                type: integer
...
    TaskEnvelope:
      description: Task payload
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/Task'
              message:
                type: string
              status:
                type: integer
    TaskListEnvelope:
      description: Task list payload
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
              message:
                type: string
              status:
                type: integer
    TaskStatsEnvelope:
      description: Task stats payload
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/TaskStats'
              message:
                type: string
              status:
                type: integer
    ConstraintEnvelope:
      description: Constraint payload
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/Constraint'
              message:
                type: string
              status:
                type: integer
    ConstraintListEnvelope:
      description: Constraint list payload
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/Constraint'
              message:
                type: string
              status:
                type: integer
    NullDataEnvelope:
      description: Command succeeded with null data
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                nullable: true
              message:
                type: string
              status:
                type: integer
...
paths:
...
  /tasks:
    get:
      tags: [task]
      summary: List tasks
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: include_suspended
          in: query
          schema:
            type: boolean
        - name: include_completed
          in: query
          schema:
            type: boolean
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/TaskListEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    post:
      tags: [task]
      summary: Create task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          $ref: '#/components/responses/TaskEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /tasks/{id}:
    get:
      tags: [task]
      summary: Get task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/TaskEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    put:
      tags: [task]
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          $ref: '#/components/responses/TaskEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    delete:
      tags: [task]
      summary: Delete task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/NullDataEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /tasks/{id}/complete:
    post:
      tags: [task]
      summary: Mark task completed
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/NullDataEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /tasks/{id}/incomplete:
    post:
      tags: [task]
      summary: Mark task incomplete
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/NullDataEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /tasks/{id}/suspend:
    post:
      tags: [task]
      summary: Suspend task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/NullDataEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /tasks/{id}/unsuspend:
    post:
      tags: [task]
      summary: Unsuspend task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/NullDataEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /tasks/stats/{date}:
    get:
      tags: [task]
      summary: Get task stats for date
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          $ref: '#/components/responses/TaskStatsEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /constraints:
    get:
      tags: [constraint]
      summary: List constraints
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/ConstraintListEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    post:
      tags: [constraint]
      summary: Create constraint
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConstraintRequest'
      responses:
        '200':
          $ref: '#/components/responses/ConstraintEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /constraints/{id}:
    get:
      tags: [constraint]
      summary: Get constraint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/ConstraintEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    put:
      tags: [constraint]
      summary: Update constraint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConstraintRequest'
      responses:
        '200':
          $ref: '#/components/responses/ConstraintEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
    delete:
      tags: [constraint]
      summary: Delete constraint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/NullDataEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /constraints/{id}/complete:
    post:
      tags: [constraint]
      summary: Complete constraint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteConstraintRequest'
      responses:
        '200':
          $ref: '#/components/responses/NullDataEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
  /constraints/{id}/reactivate:
    post:
      tags: [constraint]
      summary: Reactivate constraint
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/NullDataEnvelope'
        default:
          $ref: '#/components/responses/ApiError'
...